More generic implementation to ignore logging
This change allows to write any custom logic as to whether logging
should be done or not. By default, everything is logged, so the change
is backwards compatible. If any logic is needed to enable or disable
logging, this can be done in a custom class that implements the
`IgnoreHandlerInterface`.